LayoutTests/fast/mediastream/MediaStream-construct-with-ended-tracks.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
            var mediaStream;
            var videoTrack;
            var audioTrack;
            var newMediaStream;

            function gotStream(stream) {
                mediaStream = stream;
                debug("");
                shouldEvaluateTo('mediaStream.ended', false);

                evalAndLog("videoTrack = mediaStream.getVideoTracks()[0]");
                evalAndLog("audioTrack = mediaStream.getAudioTracks()[0]");

                shouldNotBeEqualToString('videoTrack.readyState', 'ended');
                shouldNotBeEqualToString('audioTrack.readyState', 'ended');

                videoTrack.stop();
                audioTrack.stop();
                shouldBeEqualToString('videoTrack.readyState', 'ended');
                shouldBeEqualToString('audioTrack.readyState', 'ended');

                var tracksArray = [];
                tracksArray.push(videoTrack);
                tracksArray.push(audioTrack);
                newMediaStream = new webkitMediaStream(tracksArray);

                shouldEvaluateTo('newMediaStream.ended', true);
                shouldNotBeEqualToString(mediaStream.id, newMediaStream.id);

                var mediaStreamCopy = new webkitMediaStream(newMediaStream);
                shouldEvaluateTo('mediaStreamCopy.ended', true);
                shouldNotBe('mediaStreamCopy.id', 'newMediaStream.id');
            }

            function getUserMediaError() {
                testFailed('Could not get MediaStream');
                finishJSTest();
            }

            function start() {
                try {
                    navigator.webkitGetUserMedia({audio:true, video:true}, gotStream, getUserMediaError);
                } catch (e) {
                    testFailed('webkitGetUserMedia threw exception :' + e);
                    finishJSTest();
                }
            }

            window.jsTestIsAsync = true;
            window.successfullyParsed = true;
        </script>
    </head>
    <body onload="start()">
        <p id="description"></p>
        <div id="console"></div>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
